设置了每个模型的主键，并且将类型plist文件写入了数据库

+++ MPTally/Model/MPBillModel.h
@@ -22,7 +22,7 @@
 /// 备注
 @property (nonatomic, copy) NSString *remark;
 /// 金额
-@property (nonatomic, assign) float money;
+@property (nonatomic, assign) double money;
 /// 账单类别
 @property (nonatomic, strong) MPCategoryModel *category;
 /// 账户
